Base with controllers

+++ src/modules/users/users.service.ts
@@ -1,51 +1,68 @@
-import { BadRequestException, ConflictException, Injectable, Logger, LoggerService, UnauthorizedException } from '@nestjs/common';
-import { isUndefined }                                                                                      from '@nestjs/common/utils/shared.utils';
+import { BadRequestException, ConflictException, Injectable, UnauthorizedException } from '@nestjs/common';
 
-import { EntityRepository } from '@mikro-orm/core';
 import { InjectRepository } from '@mikro-orm/nestjs';
+import { EntityRepository } from '@mikro-orm/postgresql';
 import { compare, hash }    from 'bcrypt';
+import { isInt }            from 'class-validator';
 
-import { CommonService }  from '@common/common.service';
-import { isNull }         from '@common/utils/validation.util';
-import { UserEntity }     from '@/modules/users/entities/user.entity';
-import { UpdateUserDto }  from '@/modules/users/dtos/update-user.dto';
-import { ChangeEmailDto } from '@/modules/users/dtos/change-email.dto';
+import { CommonService }       from '@common/common.service';
+import { SLUG_REGEX }          from '@common/consts/regex.const';
+import { isNull, isUndefined } from '@common/utils/validation.util';
+
+import { ChangeEmailDto } from './dtos/change-email.dto';
+import { PasswordDto }    from './dtos/password.dto';
+import { UpdateUserDto }  from './dtos/update-user.dto';
+import { UserEntity }     from './entities/user.entity';
 
 @Injectable()
 export class UsersService {
+  constructor(
+    @InjectRepository(UserEntity)
+    private readonly usersRepository: EntityRepository<UserEntity>,
+    private readonly commonService: CommonService,
+  ) {}
 
-  private readonly logger: LoggerService;
+  public async create(
+    email: string,
+    name: string,
+    password: string,
+  ): Promise<UserEntity> {
+    const formattedEmail = email.toLowerCase();
+    await this.checkEmailUniqueness(formattedEmail);
+    const formattedName = this.commonService.formatName(name);
+    const user = this.usersRepository.create({
+      email: formattedEmail,
+      name: formattedName,
+      username: await this.generateUsername(formattedName),
+      password: await hash(password, 10),
+    });
+    await this.commonService.saveEntity(this.usersRepository, user, true);
+    return user;
+  }
 
-  constructor( @InjectRepository(UserEntity) private readonly usersRepository: EntityRepository<UserEntity>,
-               private readonly commonService: CommonService ) {
-    this.logger = new Logger(UsersService.name);
+  public async findOneByIdOrUsername(
+    idOrUsername: string,
+  ): Promise<UserEntity> {
+    const parsedValue = parseInt(idOrUsername, 10);
+
+    if ( !isNaN(parsedValue) && parsedValue > 0 && isInt(parsedValue) ) {
+      return this.findOneById(parsedValue);
+    }
+
+    if (
+      idOrUsername.length < 3 ||
+      idOrUsername.length > 106 ||
+      !SLUG_REGEX.test(idOrUsername)
+    ) {
+      throw new BadRequestException('Invalid username');
+    }
+
+    return this.findOneByUsername(idOrUsername);
   }
 
   public async findOneById( id: number ): Promise<UserEntity> {
     const user = await this.usersRepository.findOne({ id });
     this.commonService.checkEntityExistence(user, 'User');
-    return user;
-  }
-
-  public async findOneByEmail( email: string ): Promise<UserEntity> {
-    const user = await this.usersRepository.findOne({
-      email: email.toLowerCase(),
-    });
-    this.throwUnauthorizedException(user);
-    return user;
-  }
-
-  public async findOneByCredentials(
-    id: number,
-    version: number,
-  ): Promise<UserEntity> {
-    const user = await this.usersRepository.findOne({ id });
-    this.throwUnauthorizedException(user);
-
-    if ( user.credentials.version !== version ) {
-      throw new UnauthorizedException('Invalid credentials');
-    }
-
     return user;
   }
 
@@ -66,22 +83,48 @@ export class UsersService {
     return user;
   }
 
-  public async create(
-    email: string,
-    name: string,
-    password: string,
-  ): Promise<UserEntity> {
-    const formattedEmail = email.toLowerCase();
-    await this.checkEmailUniqueness(formattedEmail);
-    const formattedName = this.commonService.formatName(name);
-    const user = this.usersRepository.create({
-      email: formattedEmail,
-      name: formattedName,
-      username: await this.generateUsername(formattedName),
-      password: await hash(password, 10),
+  public async findOneByEmail( email: string ): Promise<UserEntity> {
+    const user = await this.usersRepository.findOne({
+      email: email.toLowerCase(),
     });
+    this.throwUnauthorizedException(user);
+    return user;
+  }
 
-    await this.commonService.saveEntity(this.usersRepository, user, true);
+  // necessary for password reset
+  public async uncheckedUserByEmail( email: string ): Promise<UserEntity> {
+    return this.usersRepository.findOne({
+      email: email.toLowerCase(),
+    });
+  }
+
+  public async findOneByCredentials(
+    id: number,
+    version: number,
+  ): Promise<UserEntity> {
+    const user = await this.usersRepository.findOne({ id });
+    this.throwUnauthorizedException(user);
+
+    if ( user.credentials.version !== version ) {
+      throw new UnauthorizedException('Invalid credentials');
+    }
+
+    return user;
+  }
+
+  public async confirmEmail(
+    userId: number,
+    version: number,
+  ): Promise<UserEntity> {
+    const user = await this.findOneByCredentials(userId, version);
+
+    if ( user.confirmed ) {
+      throw new BadRequestException('Email already confirmed');
+    }
+
+    user.confirmed = true;
+    user.credentials.updateVersion();
+    await this.commonService.saveEntity(this.usersRepository, user);
     return user;
   }
 
@@ -107,25 +150,6 @@ export class UsersService {
       user.username = formattedUsername;
     }
 
-    await this.commonService.saveEntity(this.usersRepository, user);
-    return user;
-  }
-
-  public async updateEmail(
-    userId: number,
-    dto: ChangeEmailDto,
-  ): Promise<UserEntity> {
-    const user = await this.findOneById(userId);
-    const { email, password } = dto;
-
-    if ( !( await compare(password, user.password) ) ) {
-      throw new BadRequestException('Invalid password');
-    }
-
-    const formattedEmail = email.toLowerCase();
-    await this.checkEmailUniqueness(formattedEmail);
-    user.credentials.updateVersion();
-    user.email = formattedEmail;
     await this.commonService.saveEntity(this.usersRepository, user);
     return user;
   }
@@ -162,25 +186,38 @@ export class UsersService {
     return user;
   }
 
-  // necessary for password reset
-  public async uncheckedUserByEmail( email: string ): Promise<UserEntity> {
-    return this.usersRepository.findOne({
-      email: email.toLowerCase(),
-    });
-  }
-
-  public async remove( userId: number ): Promise<UserEntity> {
+  public async updateEmail(
+    userId: number,
+    dto: ChangeEmailDto,
+  ): Promise<UserEntity> {
     const user = await this.findOneById(userId);
-    await this.commonService.removeEntity(this.usersRepository, user);
+    const { email, password } = dto;
+
+    if ( !( await compare(password, user.password) ) ) {
+      throw new BadRequestException('Wrong password');
+    }
+
+    const formattedEmail = email.toLowerCase();
+
+    if ( user.email === formattedEmail ) {
+      throw new BadRequestException('Email should be different');
+    }
+
+    await this.checkEmailUniqueness(formattedEmail);
+    user.email = formattedEmail;
+    await this.commonService.saveEntity(this.usersRepository, user);
     return user;
   }
 
-  private async checkEmailUniqueness( email: string ): Promise<void> {
-    const count = await this.usersRepository.count({ email });
+  public async delete( userId: number, dto: PasswordDto ): Promise<UserEntity> {
+    const user = await this.findOneById(userId);
 
-    if ( count > 0 ) {
-      throw new ConflictException('Email already in use');
+    if ( !( await compare(dto.password, user.password) ) ) {
+      throw new BadRequestException('Wrong password');
     }
+
+    await this.commonService.removeEntity(this.usersRepository, user);
+    return user;
   }
 
   private async checkUsernameUniqueness( username: string ): Promise<void> {
@@ -188,6 +225,22 @@ export class UsersService {
 
     if ( count > 0 ) {
       throw new ConflictException('Username already in use');
+    }
+  }
+
+  private throwUnauthorizedException(
+    user: undefined | null | UserEntity,
+  ): void {
+    if ( isUndefined(user) || isNull(user) ) {
+      throw new UnauthorizedException('Invalid credentials');
+    }
+  }
+
+  private async checkEmailUniqueness( email: string ): Promise<void> {
+    const count = await this.usersRepository.count({ email });
+
+    if ( count > 0 ) {
+      throw new ConflictException('Email already in use');
     }
   }
 
@@ -210,11 +263,5 @@ export class UsersService {
     }
 
     return pointSlug;
-  }
-
-  private throwUnauthorizedException( user: undefined | null | UserEntity, ): void {
-    if ( isUndefined(user) || isNull(user) ) {
-      throw new UnauthorizedException('Invalid credentials');
-    }
   }
 }
